pinned

+++ Source/Bot/Commands/RemoveSystemMessages.py
@@ -4,11 +4,12 @@ from ..Bot import walle
 
 async def remove_system_msg(msg: types.Message):
     try:
-        await msg.answer(msg.reply_to_message.content_type)
-
         msg_type = msg.content_type
-    
-        if msg_type == 'pinned_message':
-            await msg.delete()
+
+        if msg:
+            if msg_type == 'pinned_message':
+                msg.delete()
+            else:
+                pass
     except Exception as e:
         print(e)
